[GTK] Add IconGtk
This is a small implementation of WebCore::Icon that holds a GIcon. It's
required for WebExtensions. There's no implementation of any paint
functions as they aren't needed yet, it's purely intended to hold the
GIcon on LInux and allow WebExtensions to return one class on all
platforms (since WebCore::Icon already has an implementation for
CocoaImage)