Implement WebExtensionURLSchemeHandler for WebKitGTK Need the bug URL (OOPS!).
Reviewed by NOBODY (OOPS!).

This creates a port of WebExtensionURLSchemeHandler for WebKitGTK. Due
to how the Cocoa version worked and how URL Scheme handlers in general function,
it was easier to create a glib-specific version than a cross-platform port.

It is impossible to test this code at the moment, the code will be tested in
a follow-up commit. If we are able to load extension-specific pages in an
Extension Context, then the scheme handler is functional.

* Source/WebKit/SourcesGTK.txt:
* Source/WebKit/UIProcess/Extensions/WebExtensionURLSchemeHandler.h:
* Source/WebKit/UIProcess/Extensions/glib/WebExtensionURLSchemeHandlerGLib.cpp: Added.
(WebKit::WebExtensionURLSchemeHandler::WebExtensionURLSchemeHandler):
(WebKit::WebExtensionURLSchemeHandler::platformStartTask):
(WebKit::WebExtensionURLSchemeHandler::platformStopTask):
(WebKit::WebExtensionURLSchemeHandler::platformTaskCompleted):